+++ example/lib/main.dart
@@ -53,14 +53,20 @@ class _MyHomePageState extends State<MyHomePage> {
       floatingActionButton: FloatingActionButton(
         onPressed: () {
           final result = _shouldIncrement(_counter);
-          if (result.isSuccess()) {
-            setState(() {
-              _counter++;
-            });
-          } else {
-            ScaffoldMessenger.of(context).showSnackBar(
-                SnackBar(content: Text("Counter cannot be bigger than 9")));
-          }
+          result.when(
+            (success) {
+              setState(() {
+                _counter++;
+              });
+            },
+            (error) {
+              ScaffoldMessenger.of(context).showSnackBar(
+                SnackBar(
+                  content: Text(error.message),
+                ),
+              );
+            },
+          );
         },
         tooltip: 'Increment',
         child: Icon(Icons.add),
@@ -68,7 +74,7 @@ class _MyHomePageState extends State<MyHomePage> {
     );
   }
 
-  Result<IncrementException, bool> _shouldIncrement(int currentCount) {
+  Result<bool, IncrementException> _shouldIncrement(int currentCount) {
     if (currentCount + 1 == 10) {
       return Error(IncrementException("counter can't be bigger than 9"));
     } else {